requirements.txt
black
pan-os-python
rich